[refactor] Project DSN 재사용을 위한 리팩토링
- 앞으로가기 뒤로가기 버튼을 제외한 모든 컴포넌트 분리
